Moving base database model to base classes directory
