+++ lib/classes/base/BaseModel.js
@@ -1,6 +1,6 @@
 const _ = require("lodash");
 const { Model } = require("sequelize");
-const { createModuleLog } = require("../../classes/Utils.js");
+const { createModuleLog } = require("../Utils.js");
 const log = createModuleLog("Database");
 
 class BaseModel extends Model {
