+++ lib/database/models/Document.js
@@ -1,4 +1,4 @@
-const BaseModel = require("./BaseModel.js");
+const BaseModel = require("../../classes/base/BaseModel.js");
 const GuildMember = require("./GuildMember.js");
 const Sequelize = require("sequelize");
