+++ lib/database/models/Guild.js
@@ -1,4 +1,4 @@
-const BaseModel = require("./BaseModel.js");
+const BaseModel = require("../../classes/base/BaseModel.js");
 const Config = require("../../classes/ConfigManager.js");
 const Sequelize = require("sequelize");
