+++ lib/database/models/GuildChannel.js
@@ -1,4 +1,4 @@
-const BaseModel = require("./BaseModel.js");
+const BaseModel = require("../../classes/base/BaseModel.js");
 const Guild = require("./Guild.js");
 const Sequelize = require("sequelize");
