+++ lib/database/models/GuildMember.js
@@ -1,4 +1,4 @@
-const BaseModel = require("./BaseModel.js");
+const BaseModel = require("../../classes/base/BaseModel.js");
 const Guild = require("./Guild.js");
 const User = require("./User.js");
 const Sequelize = require("sequelize");
