+++ lib/database/models/MessageStatistics.js
@@ -1,6 +1,6 @@
 const GuildMember = require("./GuildMember.js");
 const GuildChannel = require("./GuildChannel.js");
-const BaseModel = require("./BaseModel.js");
+const BaseModel = require("../../classes/base/BaseModel.js");
 const Sequelize = require("sequelize");
 
 class MessageStatistics extends BaseModel {
