+++ lib/database/models/User.js
@@ -1,4 +1,4 @@
-const BaseModel = require("./BaseModel.js");
+const BaseModel = require("../../classes/base/BaseModel.js");
 const Sequelize = require("sequelize");
 
 class User extends BaseModel {
