+++ lib/database/models/VoiceStatistics.js
@@ -1,5 +1,5 @@
 const GuildMember = require("./GuildMember.js");
-const BaseModel = require("./BaseModel.js");
+const BaseModel = require("../../classes/base/BaseModel.js");
 const Sequelize = require("sequelize");
 
 class VoiceStatistics extends BaseModel {
